refactor progress and add it in add to collection content component from details
