+++ presentation/details/src/main/java/com/giraffe/details/components/AddToCollectionContent.kt
@@ -1,12 +1,5 @@
 package com.giraffe.details.components
 
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
@@ -18,21 +11,16 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.drawscope.rotate
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-import com.giraffe.designsystem.theme.Theme
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.giraffe.designsystem.composable.Progress
 import com.giraffe.designsystem.theme.CineVerseTheme
+import com.giraffe.designsystem.theme.Theme
 import com.giraffe.details.R
 
 @Composable
@@ -68,50 +56,8 @@ fun AddToCollectionContent(
         }
 
         if (isLoading) {
-            AngularProgressIndicator(
+            Progress(
                 modifier = Modifier.padding(end = 12.dp),
-                color = Theme.color.brand.primary
-            )
-        }
-    }
-}
-
-@Composable
-fun AngularProgressIndicator(
-    modifier: Modifier = Modifier,
-    color: Color,
-    size: Dp = 24.dp,
-    strokeWidth: Dp = 3.dp
-) {
-    val transition = rememberInfiniteTransition()
-    val rotation by transition.animateFloat(
-        initialValue = 0f,
-        targetValue = 360f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(durationMillis = 1000, easing = LinearEasing),
-            repeatMode = RepeatMode.Restart
-        )
-    )
-
-    Canvas(modifier = modifier.size(size)) {
-        rotate(rotation) {
-            drawArc(
-                brush = Brush.sweepGradient(
-                    listOf(
-                        color.copy(alpha = 0.1f),
-                        color.copy(alpha = 0.5f),
-                        color.copy(alpha = 1f),
-                        color.copy(alpha = 0.5f),
-                        color.copy(alpha = 0.1f)
-                    )
-                ),
-                startAngle = 0f,
-                sweepAngle = 270f,
-                useCenter = false,
-                style = Stroke(
-                    width = strokeWidth.toPx(),
-                    cap = androidx.compose.ui.graphics.StrokeCap.Round
-                )
             )
         }
     }
@@ -120,7 +66,6 @@ fun AngularProgressIndicator(
 @Composable
 @Preview(
     name = "AddToCollection Light",
-    apiLevel = 34,
     showBackground = false,
     showSystemUi = false
 )
@@ -136,7 +81,6 @@ fun PreviewAddToCollectionContentLight() {
 @Composable
 @Preview(
     name = "AddToCollection Dark",
-    apiLevel = 34,
     showBackground = false,
     showSystemUi = false
 )
